libxcwm: Support for override-redirect window mapping and unmapping
Creation of override-redirect windows are done through XCB_MAP_NOTIFY,
and removal through XCB_UNMAP_NOTIFY so cases for these events have
been added to event_loop.c::run_event_loop.c

Override_redirect and inital_damage flags added to xcwm_window_t
struct.

_xcwm_window_created changed present tense _xcwm_window_create. Check
of window attributes sets override_redirect and initial_damage flags
to "1" if window is determined to be override-redirect.

Accessor function xcwm_window_is_override_redirect() added to allow
window manager to determine if window is/isn't an override-redirect
window.

_xcwm_destroy_window generalized to _xcwm_window_remove which is
called by XCB_UNMAP_NOTIFY and XCB_DESTROY_NOTIFY events.

Alteration to handing damage event to in
event_loop.c::run_event_loop().  Initial damage events to
override-redirect windows are reported relative to the root window -
not relative to the window itself. The "initial_damage" and
"override_redirect" flags in xcwm_window_t is used to determine if the
damage event is the first damage event for an override-redirect
window. If this is the case, the coordinates of the damage area of the
window are translated to window relative coordinates before the event
is passed on to the window manager client. This allows the client to
grab the correct window pixmap when it calls
xcwm_image_copy_damaged().

Signed-off-by: Jess VanDerwalker <washu@sonic.net>

Reviewed-by: Jeremy Huddleston Sequoia <jeremyhu@apple.com>