xtoq: NSFlagsChanged case added to XtoqApplication sendEvent()
Press and release of modifier keys are not passed by NSKeyUp or
NSKeyDown events, but in NSFlagsChanged. Added an NSFlagsChanged case
to the sendEvent switch.

The modifier flags of the event are checked against the current set of
modifier flags to determine if the modifier key, as specified in the
event's keyCode, has been pressed or released. Keycode is then sent on
to keyUp/DownInView handler.

Signed-off-by: Jess VanDerwalker <washu@sonic.net>

Reviewed-by: Jeremy Huddleston Sequoia <jeremyhu@apple.com>