libxcwm: Functions in atoms.c and window.c for setting of ICCCM WM_STATE atom.
A wm_state_atom field added to xcwm_atoms_t and value determined on
atom init.

Functions to change state to Iconic or Normal added to window.c and
atoms.c. Functions in window.c are wrappers for the atoms.c
functions. The window.c functions are to be called by window manager
when window is iconified or de-iconified.

Signed-off-by: Jess VanDerwalker <washu@sonic.net>

Reviewed-by: Jeremy Huddleston Sequoia <jeremyhu@apple.com>