libxcwm: Removing initial damage on override-redirect window.
The initial damage event for override-redirect windows are reported
relative to screen, not window. This initial damage is removed, and
then the whole window is marked as damaged (using window relative
coordinates).

Override-redirect can also report damage events with an incorrect
height. This case is treated the same as the initial damage event.

Signed-off-by: Jess VanDerwalker <washu@sonic.net>

Reviewed-by: Jeremy Huddleston Sequoia <jeremyhu@apple.com>