Add xcwm_window_constrain_sizing()
XtoQ appears to rely on the native windowing system to constrain the window size.
For Windows, the application has to do this itself.  Rather than marhsalling all
WM_NORMAL_HINTS data through xcwm, just add a helper function to perform the size
constraints for WM_NORMAL_HINTS

This supports  the aspect ratio constraint, which we seem to be ignoring before.

Future work: libxcwm appears to assume that WM_NORMAL_HINTS doesn't change after
the window is mapped, which doesn't seem to be guaranteed.