Get image of window with alpha channel, if available
Use xcb_composite_name_pixmap() to name the off-screen pixmap for the window,
replacing it with a new one when the window is mapped or resized.  Also name the
composite pixmap for adopted windows.

Release composite pixmap on window unmap or destroy

(We need to be more careful than currently in MAP_NOTIFY.  Always update the
name of the off-fscreen pixmap in MAP_NOTIFY, but only generate a
XCWM_EVENT_WINDOW_CREATE for override-redirect windows (since one was already
generated in MAP_REQUEST for non-override-redirect windows)

Ensure border width is zero, as border is included in composite pixmap image of
the window.