Remove unnecessary event selection and reporting
Don't select on the root window for, and report events we are not interested in.
(XCB_(KEY|BUTTON)_(PRESS|RELEASE), XCB_ENTER|LEAVE_NOTIFY)

Do select for XCB_EVENT_MASK_RESIZE_REDIRECT and XCB_EVENT_MASK_BUTTON_PRESS on
the root window, even though we are not interested in them, as only one client
is allowed to select for them, to ensure no other WM is running.