Try to fix a race in configure_request handling which causes windows to be misplaced or misized
Would this be fixed by doing xsync() after xcb_configure_window()?

Server values for values not in value mask in configure_request can be out
of date if xcb_configure_window() we have made to pass on a configure_request
hasn't completed yet...

Ignoring configure requests for windows we are managing is wrong, it means
they can't request to move themselves?