Generate XCWM_EVENT_WINDOW_CONFIGURE when a window changes size/position
Some windows may change their own size or position without a configure request
generated by us, so add an event indicating that, so native window may be
updated.

Avoid unneccesary work:

- Only update composite pixmap resources on CONFIGURE_NOTIFY when the window is
resized

- Do not send unneeded configure requests in xcwm_window_configure().  They may
cross paths with a configure request which makes them redundant.

- Only generate XCWM_EVENT_WINDOW_CONFIGURE on size/position change. Don't
generate it for stacking order or border width changes.

To do this, we need to maintain separate notified and requested window bounds:

Only note size/position changes in CONFIGURE_NOTIFY for override-redirect
windows

* notified_bounds: the most recent geometry received in a CONFIGURE_NOTIFY

We only maintain this to allow us to check for window resize and reposition, so
we know when to emit XCWM_EVENT_WINDOW_CONFIGURE events and reallocate the SHM
area.

* bounds: the most recent geometry requested.  We are authortiative for all
non-override-redirect windows, so do not allow a possibly out-of-date
CONFIGURE_NOTIFY to update this.